@@ -1,7 +1,8 @@
 import { registerLocaleData } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import localePt from '@angular/common/locales/pt';
-import { NgModule, LOCALE_ID } from '@angular/core';
+import { LOCALE_ID, NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 
@@ -11,6 +12,9 @@ import { AppComponent } from './app.component';
 import { ErrorService } from './error.service';
 import { HeaderComponent } from './header/header.component';
 import { HomeComponent } from './home/home.component';
+import { OrderItemsComponent } from './order/order-items/order-items.component';
+import { OrderComponent } from './order/order.component';
+import { OrderService } from './order/order.service';
 import { MenuItemComponent } from './restaurants/restaurant-detail/menu/menu-item/menu-item.component';
 import { MenuComponent } from './restaurants/restaurant-detail/menu/menu.component';
 import { RestaurantDetailComponent } from './restaurants/restaurant-detail/restaurant-detail.component';
@@ -20,10 +24,12 @@ import { ShoppingCartService } from './restaurants/restaurant-detail/shopping-ca
 import { RestaurantService } from './restaurants/restaurant.service';
 import { RestaurantComponent } from './restaurants/restaurant/restaurant.component';
 import { RestaurantsComponent } from './restaurants/restaurants.component';
-
-registerLocaleData(localePt);
+import { InputComponent } from './shared/input/input.component';
+import { RadioComponent } from './shared/radio/radio.component';
+import { DeliveryCostsComponent } from './order/delivery-costs/delivery-costs.component';
 
 const localeService = { provide: LOCALE_ID, useValue: 'pt-br' };
+registerLocaleData(localePt);
 
 @NgModule({
   declarations: [
@@ -37,14 +43,26 @@ const localeService = { provide: LOCALE_ID, useValue: 'pt-br' };
     MenuComponent,
     ShoppingCartComponent,
     MenuItemComponent,
-    ReviewsComponent
+    ReviewsComponent,
+    OrderComponent,
+    InputComponent,
+    RadioComponent,
+    OrderItemsComponent,
+    DeliveryCostsComponent
   ],
-  imports: [BrowserModule, RouterModule, AppRoutingModule, HttpClientModule],
+  imports: [
+    BrowserModule,
+    RouterModule,
+    AppRoutingModule,
+    HttpClientModule,
+    FormsModule
+  ],
   providers: [
     ErrorService,
     RestaurantService,
     ShoppingCartService,
-    localeService
+    localeService,
+    OrderService
   ],
   bootstrap: [AppComponent]
 })
